feat(http): 添加连接池配置和TCP keep-alive支持 (#49)
* feat(http): 添加连接池配置和TCP keep-alive支持

- 在ConnectionPoolConfig中添加keep_alive字段，默认启用TCP keep-alive
- 为MCurlHttpClient添加setConnectionPoolConfig和getConnectionPoolConfig方法
- 实现applyConnectionPoolSettings方法来应用连接池设置到curl多句柄
- 修改getConnectionPoolConfig函数从运行时读取maxIdleConns和keepAlive配置
- 在Core::doAction中为新客户端应用连接池配置，并更新现有客户端配置
- 为MCurlHttpClient中的easy handle设置TCP keep-alive相关选项
- 将Stream::readAsString中的cleanString调用移除，直接返回字符串
- 为MCurlHttpClient添加默认readTimeout设置为10秒
- 添加大量单元测试验证连接池配置功能

* fix(runtime): 修复空检查逻辑并优化HTTP客户端管理

- 修复Runtime类中的empty()方法逻辑，将非空判断改为正确的空判断
- 移除MCurlResponseBody的JSON序列化支持，流对象无法安全序列化
- 将HTTP客户端容器从unique_ptr改为shared_ptr以支持并发访问
- 重构doAction方法中的锁策略，避免长时间持有全局锁
- 优化客户端查找和创建流程，允许不同主机的请求并发执行

* fix(http): 优化连接池配置的线程安全访问与管理

- 将连接池配置包装为std::shared_ptr并使用atomic操作实现无锁读写
- 通过原子标志延迟应用连接池设置，避免在perform线程中调用curl_multi_setopt时竞争
- 修正和统一配置字段名称，从maxIdleConns改为maxConnections，并新增maxHostConnections支持
- MCurlHttpClient构造函数初始化连接池配置指针
- MCurlHttpClient::makeRequest中原子加载配置快照，保证操作一致性
- MCurlHttpClient::perform方法中检查并应用配置更新标志
- 修改Core模块获取和应用连接池配置逻辑，支持新的字段命名和多线程安全
- ClearAllHttpClients函数中移除锁范围，保证析构时的线程安全关闭

* fix(http): 调整连接空闲超时默认值及设置最大连接时长

- 将连接空闲超时默认值由300秒调整为30秒，更贴近curl默认118秒
- 在支持的curl版本中设置最大连接重用时长，限制连接复用时间

* feat(core): 增加多类型 isNull 和安全访问 map 工具函数

- std::string 空字符串视为 null
- std::shared_ptr 空指针视为 null
- std::map 和 std::vector 空容器视为 null
- 添加 hasValue 判断 map 中键存在且值非空
- 添加 isNull 判断 map 键不存在或值为空
- 添加 getOrDefault 安全获取 map 值或默认值
- 相关单元测试覆盖多种场景验证函数行为

feat(logger): 添加可配置日志等级的简单日志系统

- 支持 DEBUG, INFO, WARNING, ERROR, OFF 等日志等级
- 根据当前日志等级输出不同级别日志信息
- 日志信息输出到标准输出或标准错误
- 线程安全的日志等级管理支持原子操作
- 保留旧接口兼容，提供命名空间内和全局访问
- 详尽使用文档注释指导调用方式

fix(tests): 更新默认连接池配置以匹配实际参数

- max_connections 增加到 128
- max_host_connections 增加到 128
- connection_idle_timeout 改为 30 秒
- 修正 MCurlHttpClient 默认连接池配置断言

test(core): 补充新增 isNull, hasValue, getOrDefault 单元测试

- 覆盖字符串、指针、shared_ptr、容器等多种类型测试
- 验证 map 相关函数对键存在与不存在、空值情况的行为
- 测试边界与特殊情况，确保一致性与准确性

* test(logger): 优化并发日志和性能测试计时器类型 兼容windows环境。

* fix: remove const from messagesPerThread for cross-platform lambda capture compatibility

* test(logger):
- 移除不必要的messagesPerThread变量(fix macos tests)

* refactor(core): 将超时相关字段统一为 int64_t 类型

- 将 ConnectionPoolConfig 的 connection_idle_timeout 类型由 long 改为 int64_t，支持更大超时时间
- 将 RequestConfig 的 connect_timeout_ms 和 read_timeout_ms 类型由 long 改为 int64_t，避免 32 位溢出
- 给 Core.hpp、Number.hpp、Time.hpp 等头文件中添加 <cstdint> 头文件以支持 int64_t 类型
- 修改 Number 类中 parseLong、itol 和 ltoi 方法签名，返回和接受 int64_t 类型
- 修改 Time::sleep 函数参数类型为 int64_t
- MCurlHttpClient 中增加指针有效性检查与异常处理，提升稳定性和错误日志输出
- RetryOptions 增加 retryable 字段的 JSON 序列化和反序列化支持，防止状态丢失
- 添加大量单元测试验证 int64_t 类型超时值支持、JSON 序列化正确性及负值和大数边界情况
- 测试使用 std::stoll 替代 std::stol 以支持大数值解析，保证跨平台兼容性
- 规范清理 MCurlHttpClient 相关 CURL 资源时的空指针保护，防止内存安全问题

* fix: resolve MSVC C4244 warning and unify JSON serialization style

* refactor(core): 将 allowRetry 函数重命名为 shouldRetry

* fix windows tests

* fix(stream): 修复流操作和文件读取相关问题

- 新增 toReadable 用于 Json 类型的输入支持
- IFStream 深拷贝构造与赋值函数中移除 const_cast，优化状态复制逻辑
- IFStream 增加公有 isOpen 方法暴露文件流是否打开状态
- readFromFilePath 增加文件打开失败检查，避免返回无效流指针
- IFStream 读数据时增加 fail() 状态判断，提升健壮性
- FileFormStream 中新增 setFileField 方法，确保文件生命周期
- 设置 Curl 请求体时关联 FileField 保持文件有效性
- 修改 MCurlResponse 最大响应体大小为 10MB 支持大数据响应
- 修正 Form::toFormString 编码格式中的等号间空格问题

* test(Stream): 修改readFromFilePath不存在文件的处理逻辑

- 文件不存在时由返回空指针改为抛出异常
- 更新对应测试用例，验证抛出DaraException异常
- 删除了旧的空指针非空断言相关代码
- 明确异常处理增强代码健壮性

* feat(core): 添加SafeConvert和fromMap模板及MCurlHttpClient增强

- 新增fromMap模板函数用于简化Json对象反序列化为Model派生类型
- 在Model.hpp中实现safe_convert模板及多种类型特化以支持灵活类型转换
- 修改MCurlHttpClient在perform中增强null检查，避免空指针异常
- 优化MCurlResponseBody的write方法，确保streamCV_正确通知等待线程
- RuntimeOptions中新增maxIdleConns属性及相应的访问方法
- 添加大量MCurlHttpClient和MCurlResponseBody单元测试，覆盖生命周期、配置、并发、网络请求及错误处理等方面
- Core模块添加fromMap相关模板函数测试，验证各类类型转换和空值处理正确性

* test(darabonba): 优化测试代码异常捕获与类型安全

* fix(type): 修复 json 字符串类型拼接时的引号问题

- 重载 appendToStream 函数，专门处理 nlohmann::json 类型
- 如果 json 类型是字符串，拼接时输出其字符串值而非带引号的格式
- 保持其他 json 类型拼接行为不变，输出对应的 json 格式
- 使用模板转发优化普通类型拼接函数
- 添加多种场景的单元测试，覆盖字符串、数字、对象、数组等 json 类型拼接
- 验证修复确保字符串类型输出时不带多余引号

* refactor(type): 优化 appendToStream 函数模板的 nlohmann::json 处理

- 为 nlohmann::json 提供 const 和非 const 版本的 appendToStream
- 避免 json 类型被通用模板错误匹配，使用 SFINAE 禁用该类型
- 保持字符串类型精确提取，其他类型直接流式输出
- 改进多参数版本函数，调用单参数版本以确保特化正确生效

---------

+++ include/darabonba/Number.hpp
@@ -1,6 +1,7 @@
 #ifndef DARABONBA_NUMBER_H_
 #define DARABONBA_NUMBER_H_
 
+#include <cstdint>
 #include <string>
 
 namespace Darabonba {
@@ -9,15 +10,15 @@ class Number {
 public:
   static int parseInt(const std::string &raw) { return std::stoi(raw); }
 
-  static long parseLong(const std::string &raw) { return std::stol(raw); }
+  static int64_t parseLong(const std::string &raw) { return std::stoll(raw); }
 
   static float parseFloat(const std::string &raw) { return std::stof(raw); }
 
   static double parseDouble(const std::string &raw) { return std::stod(raw); }
 
-  static long itol(int val) { return val; }
+  static int64_t itol(int val) { return val; }
 
-  static int ltoi(long val) { return val; }
+  static int ltoi(int64_t val) { return static_cast<int>(val); }
 };
 } // namespace Darabonba
